+++ push_swap/libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: adurusoy <adurusoy@student.42.fr>          +#+  +:+       +#+        */
+/*   By: adurusoy <adurusoy@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/19 20:21:23 by adurusoy          #+#    #+#             */
-/*   Updated: 2023/08/19 20:36:12 by adurusoy         ###   ########.fr       */
+/*   Updated: 2023/08/20 04:42:55 by adurusoy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@
 # include <stdlib.h>
 # include "../ft_printf/ft_printf.h"
 
-int		ft_atoi(const char *str);
-void	*ft_calloc(size_t count, size_t size);
-char	**ft_split(char const *s, char c);
-char	*ft_strchr(const char *s, int c);
-char	*ft_strjoin(char const *s1, char const *s2);
-int		ft_strlen(char *a);
-int		ft_strncmp(const char *s1, const char *s2, size_t n);
+long long int	ft_atoi(const char *str);
+void			*ft_calloc(size_t count, size_t size);
+char			**ft_split(char const *s, char c);
+char			*ft_strchr(const char *s, int c);
+char			*ft_strjoin(char const *s1, char const *s2);
+int				ft_strlen(char *a);
+int				ft_strncmp(const char *s1, const char *s2, size_t n);
 
 #endif
